apps/streamlit_apps/requirements.txt
streamlit
requests
pandas
beautifulsoup4
python-dotenv
google-generativeai
scikit-learn
numpy
openpyxl
plotly